# Conflicts:
#	api/src/main/java/tpo/uade/api/service/IMyProfileService.java
#	api/src/main/java/tpo/uade/api/service/implementation/MyProfileService.java

+++ api/src/main/java/tpo/uade/api/repository/NavigationRespository.java
@@ -2,13 +2,12 @@ package tpo.uade.api.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import tpo.uade.api.model.NavigationModel;
-import tpo.uade.api.model.ProductModel;
 
 import java.util.List;
 
-
 @Repository
-public interface NavigationRespository extends JpaRepository<NavigationModel, Integer> {
+public interface NavigationRespository extends JpaRepository<NavigationModel, Long> {
     List<NavigationModel> findByUser_UserId(Long userId);
 }
